% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_info_extractor.R
\name{GOF_glm_info_extractor}
\alias{GOF_glm_info_extractor}
\title{Implements the "interface" GOF_model_info_extractor for
  for generalized linear models}
\description{
This class is specialized in extracting various information
  from an object of class "glm"
}
\section{Super class}{
\code{\link[bootGOF:GOF_model_info_extractor]{bootGOF::GOF_model_info_extractor}} -> \code{GOF_glm_info_extractor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-yhat}{\code{GOF_glm_info_extractor$yhat()}}
\item \href{#method-y_minus_yhat}{\code{GOF_glm_info_extractor$y_minus_yhat()}}
\item \href{#method-beta_x_covariates}{\code{GOF_glm_info_extractor$beta_x_covariates()}}
\item \href{#method-clone}{\code{GOF_glm_info_extractor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-yhat"></a>}}
\if{latex}{\out{\hypertarget{method-yhat}{}}}
\subsection{Method \code{yhat()}}{
see \link{GOF_model_info_extractor}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_info_extractor$yhat(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_info_extractor}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_info_extractor}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-y_minus_yhat"></a>}}
\if{latex}{\out{\hypertarget{method-y_minus_yhat}{}}}
\subsection{Method \code{y_minus_yhat()}}{
see \link{GOF_model_info_extractor}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_info_extractor$y_minus_yhat(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_info_extractor}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_info_extractor}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-beta_x_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-beta_x_covariates}{}}}
\subsection{Method \code{beta_x_covariates()}}{
see \link{GOF_model_info_extractor}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_info_extractor$beta_x_covariates(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_info_extractor}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_info_extractor}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_info_extractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
