% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_trainer.R
\name{GOF_glm_trainer}
\alias{GOF_glm_trainer}
\title{Implements the "interface" GOF_model_trainer for
  for generalized linear models}
\description{
refits an object of class "glm" to a new data set
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GOF_glm_trainer-refit}{\code{GOF_glm_trainer$refit()}}
\item \href{#method-GOF_glm_trainer-clone}{\code{GOF_glm_trainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOF_glm_trainer-refit"></a>}}
\if{latex}{\out{\hypertarget{method-GOF_glm_trainer-refit}{}}}
\subsection{Method \code{refit()}}{
see \link{GOF_model_trainer}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_trainer$refit(model, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_trainer}}

\item{\code{data}}{see \link{GOF_model_trainer}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_trainer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOF_glm_trainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GOF_glm_trainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_glm_trainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
