% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_simulator.R
\name{GOF_lm_sim_wild_rademacher}
\alias{GOF_lm_sim_wild_rademacher}
\title{Implements the "interface" GOF_model_simulator for
  for linear models}
\description{
Rademacher random variables are used to change
add or substract the residuals from the fitted values
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-resample_y}{\code{GOF_lm_sim_wild_rademacher$resample_y()}}
\item \href{#method-clone}{\code{GOF_lm_sim_wild_rademacher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resample_y"></a>}}
\if{latex}{\out{\hypertarget{method-resample_y}{}}}
\subsection{Method \code{resample_y()}}{
a wild bootstrap using Rademacher random
  variables to resample the dependent variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_lm_sim_wild_rademacher$resample_y(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_simulator}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_simulator}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_lm_sim_wild_rademacher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
