% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_simulator.R
\name{GOF_model_simulator}
\alias{GOF_model_simulator}
\title{R6 Class representing a generator/resample of the dependent variable}
\description{
R6 does not offer interfaces. Hence all methods
  are considered as abstract.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-resample_y}{\code{GOF_model_simulator$resample_y()}}
\item \href{#method-clone}{\code{GOF_model_simulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resample_y"></a>}}
\if{latex}{\out{\hypertarget{method-resample_y}{}}}
\subsection{Method \code{resample_y()}}{
Abstract function that resamples/generates
  the dependent variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_model_simulator$resample_y(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{fitted model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generates the dependent variable according to
  the \code{model}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_model_simulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
