% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_test_injector.R
\name{GOF_model_test_injector}
\alias{GOF_model_test_injector}
\title{Creates a GOF-model-test-class}
\usage{
GOF_model_test_injector(
  model,
  data,
  nmb_boot_samples,
  simulator_type,
  y_name,
  Rn1_statistic,
  gof_model_resample_class,
  gof_model_test_class
)
}
\arguments{
\item{model}{see \link{GOF_model_test}}

\item{data}{see \link{GOF_model_test}}

\item{nmb_boot_samples}{see \link{GOF_model_test}}

\item{simulator_type}{see \link{GOF_model}}

\item{y_name}{see \link{GOF_model_test}}

\item{Rn1_statistic}{see \link{GOF_model_test}}

\item{gof_model_resample_class}{\link{GOF_model_resample} class}

\item{gof_model_test_class}{\link{GOF_model_test} class}
}
\value{
instance of \link{GOF_model_test}
}
\description{
Simplifies the creation of a GOF-model-test-class
}
