% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_trainer.R
\name{GOF_model_trainer}
\alias{GOF_model_trainer}
\title{R6 Class representing a trainer for fitting models}
\description{
R6 does not offer interfaces. Hence all methods
are considered as abstract.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-refit}{\code{GOF_model_trainer$refit()}}
\item \href{#method-clone}{\code{GOF_model_trainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refit"></a>}}
\if{latex}{\out{\hypertarget{method-refit}{}}}
\subsection{Method \code{refit()}}{
Abstract function refits the model to
  a new data set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_model_trainer$refit(model, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{fitted model}

\item{\code{data}}{used for refitting the model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{model} refitted on \code{data}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_model_trainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
