% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_simulator.R
\name{GOF_sim_wild_rademacher}
\alias{GOF_sim_wild_rademacher}
\title{Implements the "interface" GOF_model_simulator
  in a semi-parametric fashion}
\description{
This is a model agnostic resampling class, where
Rademacher random variables are used to add or substract
the residuals from the fitted values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GOF_sim_wild_rademacher$new()}}
\item \href{#method-resample_y}{\code{GOF_sim_wild_rademacher$resample_y()}}
\item \href{#method-clone}{\code{GOF_sim_wild_rademacher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_sim_wild_rademacher$new(gof_model_info_extractor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gof_model_info_extractor}}{the info extractor that is used
to derive the residuals and fitted values for resampling.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resample_y"></a>}}
\if{latex}{\out{\hypertarget{method-resample_y}{}}}
\subsection{Method \code{resample_y()}}{
a wild bootstrap using Rademacher random
  variables to resample the dependent variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_sim_wild_rademacher$resample_y(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{see \link{GOF_model_simulator}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{GOF_model_simulator}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOF_sim_wild_rademacher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
