% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-example.R
\name{not_covered}
\alias{not_covered}
\title{example of a function that is not covered by unit-tests}
\usage{
not_covered()
}
\value{
no explicit return value
}
\description{
example of a function that is not covered by unit-tests
}
\examples{
not_covered()
}
