% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rn1_statistic.R
\name{Rn1_KS}
\alias{Rn1_KS}
\title{Kolmogorov-Smirnov-statistic for marked empirical
  process}
\description{
Implements the "interface" Rn1_statistic
}
\section{Super class}{
\code{\link[bootGOF:Rn1_statistic]{bootGOF::Rn1_statistic}} -> \code{Rn1_KS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calc_statistic}{\code{Rn1_KS$calc_statistic()}}
\item \href{#method-clone}{\code{Rn1_KS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_statistic"></a>}}
\if{latex}{\out{\hypertarget{method-calc_statistic}{}}}
\subsection{Method \code{calc_statistic()}}{
calculates the Kolmogorov-Smirnov-statistic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rn1_KS$calc_statistic(Rn1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rn1}}{see \link{Rn1_statistic}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
see \link{Rn1_statistic}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rn1_KS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
