% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rn1_statistic.R
\name{Rn1_statistic}
\alias{Rn1_statistic}
\title{R6 Class representing statistics for marked
  empirical processes}
\description{
R6 does not offer interfaces. Hence all methods
  are considered as abstract.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calc_statistic}{\code{Rn1_statistic$calc_statistic()}}
\item \href{#method-clone}{\code{Rn1_statistic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_statistic"></a>}}
\if{latex}{\out{\hypertarget{method-calc_statistic}{}}}
\subsection{Method \code{calc_statistic()}}{
Abstract function that calculates the statistic
  for a given marked empirical process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rn1_statistic$calc_statistic(Rn1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rn1}}{marked empirical process as a double vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
statistic based on \code{Rn1}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rn1_statistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
