% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_simulator.R
\name{rrademacher}
\alias{rrademacher}
\title{Generates Rademacher distributed random variables}
\usage{
rrademacher(n)
}
\arguments{
\item{n}{number of random variables to be generated}
}
\value{
vector of values following the Rademacher distribution
}
\description{
Generates Rademacher distributed random variables
}
