% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_model_test_injector.R
\name{GOF_model}
\alias{GOF_model}
\title{Creates a GOF-model-test-class}
\usage{
GOF_model(model, data, nmb_boot_samples, simulator_type, y_name, Rn1_statistic)
}
\arguments{
\item{model}{see \link{GOF_model_test}}

\item{data}{see \link{GOF_model_test}}

\item{nmb_boot_samples}{see \link{GOF_model_test}}

\item{simulator_type}{either "parameteric" or "semi_parameteric_rademacher"}

\item{y_name}{see \link{GOF_model_test}}

\item{Rn1_statistic}{see \link{GOF_model_test}}
}
\value{
instance of \link{GOF_model_test}
}
\description{
Simplifies the creation of an instance of
  \link{GOF_model_test}
}
